feat: add optional SSE middleware to obtain Grafana info from headers
This PR adds a flag to the CLI which, when combined with the
SSE transport, allows the client to include information about
the Grafana instance to use in headers.

This is a non-standard way to use the MCP server, but I'm not sure any
of the other accepted auth proposals meet our use case.

Unfortunately it involves changing some of the callsites because the
Grafana client is no longer a singleton (since it can be changed).
Instead, settings live in a contextvar and the client is recreated
for each request.